End study if user turns on private browsing autostart. Follow up for #147.
In PR #149, I added an eligibility requirement that this private browsing autostart pref is not enabled. However, that does not end the study if the user turns the pref on during the study. This PR handles that case.

I added this pref observer to bootstrap.js rather than Feature.jsm, since it really should be a check all Shield studies make generally, just like study expiry.
I also:
* Updated the Study Endings section in the README
* Added this new ending to Config.jsm so the user gets a survey
* Reformatted the first arg into Feature.endStudy to be an object with a 'reason' key to better match StudyUtils.endStudy, for which it is a wrapper. I verified that each of these calls to endStudy still work as expected.