Dismiss panel on window resize. Fixes #173. Fixes #152.
After having all kinds of problems on Windows to display the intro panel correctly on window resize, I decided to change the behavior of the intro panel (and pageAction panel) to be dismissed for all platforms when the window resizes, to ensure consistent, non-buggy behavior. I checked this change with pdol, the stakeholder.