Correctly remove pageAction button wrapper element from browser.xul. Fixes #178.
Inspecting the pageAction element on the page, it looks like the wrapper element I added to get the badge to display green in #151, 'hbox' id=”tracking-protection-study-button-wrapper”, is getting added anew every time a new tab is opened. Since it has a CSS style of 'margin-right: 3px', each time a new box is added, it’s 3px to the left of the old box, thus the shifting to the left of the other, built-in pageAction icons.

In the 'Feature.hidePageAction' method, I removed the pageAction button element from the 'browser.xul' document, but not its wrapper 'hbox' element. This patch correctly removes the wrapper element (which consequently removes the pageAction button 'toolbarbutton' XUL element as well.