Ensure fast, private UI surfaces on restart. Fixes #193.
I’ve noticed with rhelmer that this does NOT happen in 60, but it does happen in 59. In 59, on restart, there’s a console error in the browser console `TypeError: win is null[Learn More]  Feature.jsm:1068:21`. This error is not present in 60, which made me think it is related to the issue.

Bootstrap.js is set up such that Feature.jsm should be gated on there being a window available, since in bootstrap.js, the feature doesn’t start up until either a window is returned by 'Services.wm.getMostRecentWindow(“navigator:browser”)' or the 'sessionstore_window_restored' event fires. While I now use the 'RecentWindow.jsm' module in my add-on inside Feature.jsm after fixing #139 to get the topmost window (which is what is returning 'null' on Firefox startup), I noticed I use 'Services.wm.getMostRecentWindow("navigator:browser")' in bootstrap.js to gate Feature.jsm (which is not returning 'null' on Firefox startup). Clearly the former appears to be filtering out other kinds of windows that we may not be interested in, while the latter does not. I need to ask in IRC #fx-team for more details, but @rhelmer and I have verified that this fixes the problem reported by SoftVision. For now, I've simply made the way in which I determine the first window consistent by replacing 'Services.wm.getMostRecentWindow(..)' with 'RecentWindow.getMostRecentBrowserWindow(..)'.